Fixed weird problem with Org and removed stuff
Whatever was happening was causing a new node with Org Roam to have a
weird title, something like:

title: #+title: Example Title

Which I could not fix, but only partially identify. Aparently it was
either the enabled spell +flyspell packages and/or the +hugo on org.
Also, there was a possible error in the org line in init.el, it looked
kind of like this:
       (org +roam2 + hugo + pretty)               ; organize your plain life in plain text
Which I changed to:
       (org +roam2 +hugo +pretty)               ; organize your plain life in plain text

Then reloaded the configuration and now I even have keybinds for org
roam. Still need to try (later) to get spell working.

Also removed the Gnome session configuration (which was maybe not even
important and could have been set somewhere else).

Commented out emms configuration for now. I just noticed it is available
in the :app section, so if I enable it, I'll probably get better
configuration by default by DOOM Emacs